urls fix

+++ src/logic/radio.py
@@ -14,26 +14,33 @@ COVERS_URL = "https://raw.githubusercontent.com/Gild56/gild56_website_lists/refs
 
 
 def get_duration(url: str) -> float:
-    with tempfile.NamedTemporaryFile(suffix=".mp3", delete=False) as temp:
+    with tempfile.NamedTemporaryFile(
+        suffix=".mp3",
+        delete=False
+    ) as temp:
         temp_path = temp.name
 
     try:
         urllib.request.urlretrieve(url, temp_path)
         return MP3(temp_path).info.length
     finally:
-        os.remove(temp_path)
+        if os.path.exists(temp_path):
+            os.remove(temp_path)
 
 
 def get_song(song: str) -> str:
-    return f"{MUSIC_URL}/{quote(song)}"
+    return f"{MUSIC_URL}/{quote(song, safe='')}.mp3"
 
 
 def get_cover(author: str) -> str:
-    return f"{COVERS_URL}/{quote(author)}.png"
+    return f"{COVERS_URL}/{quote(author, safe='')}.png"
 
 
 def get_all_songs():
-    url = "https://api.github.com/repos/Gild56/gild56_website_lists/contents/music"
+    url = (
+        "https://api.github.com/repos/"
+        "Gild56/gild56_website_lists/contents/music"
+    )
 
     token = os.getenv("GITHUB_TOKEN")
 
@@ -44,7 +51,10 @@ def get_all_songs():
     if token:
         headers["Authorization"] = f"Bearer {token}"
 
-    request = urllib.request.Request(url, headers=headers)
+    request = urllib.request.Request(
+        url,
+        headers=headers
+    )
 
     with urllib.request.urlopen(request) as response:
         data = json.load(response)
@@ -52,7 +62,8 @@ def get_all_songs():
     return [
         file["name"].removesuffix(".mp3")
         for file in data
-        if file["name"].endswith(".mp3")
+        if file["type"] == "file"
+        and file["name"].endswith(".mp3")
     ]
 
 
@@ -85,10 +96,8 @@ class RadioScheduler:
 
             return start + pause
 
-        # URL GitHub
         web_path = get_song(path)
 
-        # Durée du MP3
         duration = get_duration(web_path)
 
         self.schedule.append({
